FCOM-1 Don't error if path option is a deleted file
I had erroneously thought that `-p` in the git command stands for
"path". It actually stands for `--patch`!, and the `-p`/`--patch` option
has no relationship to the file path, as I had previously thought.

Therefore, it's no problem to separate out the file path after a ` -- `
separator in the command, which has the advantage of causing the git
command not to error if a file has been deleted.