Default `--rg-options` to `--max-columns=2000` [FCOM-3]
I rarely actually have interest in looking at a line that is more than
2,000 characters long. Usually, I think that such a line should never
have been committed to the repo, in the first place. So, I think it's
pretty rare for me to actually care about such a line. One exception
might be an unwrapped line in a markdown file, but hopefully a 2,000
character limit won't eliminate many such lines.