Add '--topo-order' (topological order) option to 'git log' command
This will order commits as per their order of application to the
codebase, rather than the chronological order of the commit dates. These
orderings can differ, particularly when it comes to merged branches. I
think that generally what we want when they differ is the topological
order, rather than the chronological order.