Stream results progressively [FCOM-8] (#757)
This change brings back streaming of results, which I think was broken
in #639 .

Unsurprisingly, given the complexity/overhead added to the
implementation, the fix herein does hurt performance somewhat, but not
terribly. I tested with the following queries in `david_runger` and got:

1. **with latest release:** `time fcom javascript` => 12.9s, 12.7s
1. **with this change:** `time /home/david/code/fcom/exe/fcom
javascript` => 15.1s, 14.6s

I think that this is well worthwhile, in order to be able to start
seeing streaming results as soon as they begin to become available
(which often allows for Ctrl-C to interrupt the command before it
finishes, anyway, resulting in a net time savings, even if it is slower
if/when run all the way to completion).

Private reference link:
https://chat.deepseek.com/a/chat/s/6eb15fc5-626f-48b9-b329-a0fd6933d96a