Use git '--diff-algorithm=default' (#841)
https://git-scm.com/docs/diff-options#Documentation/diff-options.txt-code--diff-algorithmpatienceminimalhistogrammyerscode

**Motivation:** For one thing, it sounds like the `default` (`myers`)
algorithm is probably faster than some of the options (which a user
might have configured in their git config), so this might result in a
performance improvement.

More importantly, it seems that the other tooling that I use (the VS
Code GitLens git blame, GitHub diffs) use the `default` diff algorithm,
and so using that diff algorithm produces results that better align with
those tools.

As a concrete illustration, in the `david_runger` repo at bc3dd3e9,
executing `/home/david/code/fcom/exe/fcom 'moduleResolution' -p
tsconfig.json --development` prior to this change would show just one
commit, 93729389. However, with this change, that same command will show
a later commit, efdc0b36, which (when using the `default` git diff
algorithm, but not when using the `histogram` algorithm) appears to have
moved that line. When viewing efdc0b36 in GitHub, it shows that the line
moved. Also, when putting the cursor on that line in VS Code, GitLens
blamed efdc0b36, not 93729389, which, prior to this change, `fcom` would
seem to point to.